Styling Additions
I aligned the text-boxes a little better, added some opacity to the placeholder text, added a forgot password link, added a basic animation when you click on the inputs. For now, the line expands only to the left but I want to make it expand in both directions as that would look a lot better.