New Burger Menu update #1
Im really pinning my hopes on the fact that we can revert if this fails. But I completely removed the old Burger menu, and im adding the new one. This is going to take a little so its going to need to be split into a few different commits. So here is the first one, just avoid editing the navbar files until the is done.