Made About Route and added profile option in Navbar
Im going to need your help here, but what needs to happen is when the home page is loaded, it needs to check if your logged in. and if you are logged in, then it shoudlnt show the login and sgnup links, and it will show the profile link. For now I just left all of them in at once, but eventully it should be like that.