Remaking NavBar functionality
I wasent fully happy with the navbar shrink and resize functionality, because it would be complicated to add or remove navbar links withough doing a lot on css. So I developed the biggest brain js script that allows us to add and remove the navbar links as we please and the resizing will always work fine. Its not fully done, I still have to make the more dropdown show again, but im done for today. (I can explain the js script if you want)