new: navbar links change based on user login state + displaying them with for loop
IMPORTANT: this is still version 1. I had reloading issues if I kept the links in the Navbar component. Basically I was using v-if and there was a really quick flash of links such as login and signup even tho the user was logged in. So I moved everything inside a separeted module and I had to use window.location instead of .push inside the profile component otherwise it wouldn't reload the navbar links (profile was still present after logout). I'm planning on working on that method and fixing it because I don't like window.location. In case it wasn't clear our checkIfLoggedIn function needs to reach to the API and the navbar content depends on the result of that, so we need a way to delay the displayin g of the links based on that result. I am thinking of a transition which just delays the rendering.