fix: switched to a much more clean way to display navbar links + organized stuff
Sorry if this is a big commit but I finally managed to nail it after a morning of pain. Now we are using an EventBus (Vue thing) to broadcast the login state of the user between components, this way we can avoid using that window.location I was talking about before. You will find comments inside the code, if you have any questions hit me up. By the way we can use the EventBus for other stuff too.