Started making the navbar from the mockups
Carefull looking at it for too long because you will lose 75% of your eyesight. But I colored it like this because I have to go but am not done, so this is just a start. I basically just restructued the css and html a bit, hopefully in the future we can re-add the NavBarLinks file, but for now this is a start