new: login and signup links show up only if logged out and profile link does the opposite
I did something sketchy here. I have noticed a bug present since my commit on NavbarLinks (the one that was apparently working for the signup, login and profile thing): if you refreshed the page the login and singup links showed up even tho you were logged in and the profile link was hidden. So to get consistency now we set a 'loggedin' cookie which is basically a flag that tells so every time we create the Navbar. The cookie is Base64 encoded to add a very subtle layer of security, so people don't see 'loggedin=false' and easily change that. If they still manage to do so the bug is very subtle too, the links will show up but as soon as they move around they switch back based on the login state. If you don't like this hacky thing we could think of ASYNC components, so basically rendering the navbar only after the server response to check if a user is logged in but that may cause delays and stuff