Revamped Dropdown Component
Before I would just use a bunch of props to get enough data to make a dropdown, but not using slots I made it even better. The only two props are spacing and label. Height and width are decided by the local style of the prop. All the links are added inside the Dropdown tag inside <li> tags.