Added dynamic routing to settings route
I made route params to decide what section of the settings page you want to get (account, display, secuirty, etc), however I think we need to make some changes going forward. First of all, I think we need a more concrete method of checking if the user is logged in on every page load and being able to wait to load the page until we get that data, and second I think we need to change our settings load in the sense that instead of getting the settings and loading them in a v-for loop, I think we should hard code in the settings and maybe just fetch values from the server that we need. I was messing around with the settings.js page and I think its going to make things more complicated than to just put the values in the settings pages. That being said, we could make more settings componenets for each setting subsection (account, secuirty, etc), and then depending on the parameter in the url, it would render one of those components in the settings_selection container. Hope this makes sense :)