Example Routing for settings
I explained a lot on discord, but if you go to the settings page, and look at the url, it will have the seciton of settings you request as the last item. So if you manually change that to something like 'privacy' or 'appearance' you will see how it works. The idea is that those links will just add a new param to the url, this way we dont need settings.js. Like I said it was getting complicated really fast, and since we would need new compoenents for every different setting section anyway I figured it was a pretty clean solution. Now we just need to figure out how to make the buttons on the side actully change the url to the corresponding section, which is diffuicult right now because its being rendered with a v-for. Which dpening on what you think we might just rid of. Hope this makes some sense :')