new: making use of dynamic components
This is a first look at what I was thinking, as you can see we dynamically attach settings options components based on a click event on the options in the left side. This way we don't have to hard code all the components and we don't even need if statements. You will find comments in the code. We also keep track of the previously selected item (I am not doing stuff with it). We just need a way to dynamically change the URL based on what the user selects.