Small Stuff
Moved the @click action on the <p> tag to the <li> in the v-for loop so you can click anywhere in the box to register it. Also, I started making the settings for appearance, but I decided to wait because we need to figure two things out: 1) How to import a stylesheet into the components 2) how to globally register a component. If you could help me with those two it would be great!