NavBar Restructure Completed
The only thing I dont quite like about the functionaity at this point, is when your on the settings screen, if you change settings tabs the burger menu stays open. We need to find a way to close the burger menu if you click anywhere outside of it. If you could help me figure this out it would be helpfull.