fix: fixed issues caused by external styles sheet
Pretty big commit actually. External sheets where applied globally and we didn't want that. Wrappers and slots now solve this problem because we register scoped css that gets applied to whatever we pass to the Wrappers. The things we want to pass are the settings components such as Account.vue, Privacy.vue etc