edit: tried to redesign burger menu
i still have mixed feelings about this, i like how it looks but i don't know if it fits well with the overall app. Also, the color inside the theme switcher are the same for both themes, I left that because we may want different colors. I thought this color suited both themes