fix: the watcher is now called once
Now we have a SettingList component which contains the SettingListItems. The watcher is located in the parent component and it passes a prop that identifies if a setting label is active. This way we avoid repeating the watcher for all the SettingListItems and we keep the code cleaner even tho it may be a little bit more complicated