Style Changes and New Route
The settings icons now change color with the theme, and I added a router for notifications. I hestiated on adding the routes for projects and market because I was thinking maybe they would be the same thing. Like when you view projects it shows all of them despite price. Im not sure though so this is something we need to talk about.