fix: directive import was messed up
There was no default export + you were registering the directive as global, which defeats the purpose of exporting the directive to import it only in the navbar. Plus names were confusing. Now we export the directive and we use it locally in the navbar, we could also make a mixin for directives but since this one is going to be used only for the navbar it defeats the purpose of using a mixin. The directive still doesn't work but at least now we can work on it