Fixed Burger Menu
The burger menu now works perfectly with fixed postitioning. I had to pull some JS powermoves to make it work, and I did something a bit strange. To make the x on the burger menu not get pulled out of the flow when I gave it a position static, I just added a placeholder div that gets toggled to make it appear like everything is fine.