Started Profile Page & Z-index Fix
The navbar was behaving strangely because it had a parent container with a fixed position. This meant all the other burger menu elements would sometimes be limited to a z-index of zero. This caused a ton of layering problems, but its fixed now. I also started a bit on the profile page.