new: users profile system
I didn't get rid of /profile because I want you to see this first. It works well in my opinion and since we will have something like /user/:username to visit users profiles why don't we use this system even for the current logged in user. So now if you are logged in as 'jack' and you visit 'jack' it will display the profile page you were working on, otherwise it will display another thing which in the future will be the first mobile mockup on the left. Check the code and the comments