fix: login cookie / navbar links bug
I have managed to fix it, the problem was caused by other cookies: the js method document.cookie returns a long ass string of all the cookies so we were trying to decode the whole thing as if it was only the cookie for the login state. So since js doesn't have a native cookie parser (it is a pain to do it from scratch trust me) I have installed a very light weight cookie parser for Vuejs, created a cookies service and handled everything with that. With this module you can literally go 'getCookie(cookieName)' and it parses it for you in a js object. If oyu have any questions about the code hit me up, it should be easy tho. For now I left the cookie in plain text instead of encoding it with Base64.