Username shown in Profile
The problem here is I ask the server for username twice, because I cant do the url check in user.vue unless the username is already established. So we need to do some async crap. Also we need to figure out a little better how we want to structure the user data being stored in vuex. Check out the changes and I think you will understand what I mean