Added Dynamic Project Cards (READ DESC)
Ait so this is a rather big commit, I coded the front-end for the project cards with placeholder project images. Right now, the cards a re added dynamically in the mobile/profilesections components with some Array of objects to make it easy to make then. I hard coded some examples for now, but eventully we need to fetch these from the API.