Update to profile page routing (Read Desc!)
So Im going to raise soeme concers I have with how we setup our components. We could get it to work, but I realized that if you are on desktop view say at /user/jacopo/projects, and then you resize the screen the url stays the same but it loads another component. Its setup the same way with the settings page. As much as I hate to say it I think we over did something here. I think it might make more sense to have the routing be a little more directly related to the actual pages being loaded. Because rn, when you load the user page, you actully look at url to ssee which componenets to render. This might be fine, but im not sure. My point basically is its getting complicated when you have a bunch of different sub routes you need to handle because for each one were going to need a mobile view and a dekstop view. MY suggestion is to either make one compoenent that handles both (with media queries) but this might get complicated. Or we make a component for every single subroute on both mobile and desktop view. I think while were still generally in the early stages we should re-consider how were setting up this page routing for things inside pages. Like the settings page and now the user page. For now I did it how we did on settings page but as I said if were gonna make it work fully were gonna need to make so many components. Let me know what you think because I honestly dont really know here.