Basic Nested Route Implementation
Ok so even if we dont eliminate the need for all the different components, this makes routing between them so easy! Im going to add it to the settings page eventully because it eliminates the need for any watchers, route pushing, etc. Before, we kinda just gave it the illusion that we were routing something and would check the route params to see if it contained the name of the component and conditionally render it. Now we literally route to the links and for now im using a conditional class check to make it styled differently if the url contains the component name but there is no reason we couldnt use a router-link-active style or something. I posted some info on this in discord but let me know what you think, and maybe you could even improve it further!