Started on working graphql endpoint for user (DESC)
So it wont work yet. Im trying to get the result of an async function in the resolve block but I just can get it work. If I could get a second pair of eyes it would be nice. Iv tried returning the actual promise, and returning the value of the promise but nothing is working. I left a bunch of comments if they help at all