Ok, This is Epic (DESC AGAIN)
AIT, this is clean af. So basically I was trying to resolve a resolved promise or something. So I literally make a new promise now in the getUserEntry() function. Check it out in the localhost:5000/graphql page because we can now just selectively retrieve only the data that we want really easily. Let me know what you think, and sorry if it sounds like im high, im a bit tired.