Big Progress on project feed (DESC)
So I did a bunch of frontend, and I started adding some of the backend functionality. However I came across a problem: I wanted to use graphql to query for the username, but I realized that to fetch user details you need the username -_-  So I used the userservice method that uses 'fetch', So im wondering if we just use this as a base plate to get the username and then use graphql for everything else. Or if there is some other way to identify users and then use graphql. I think keeping the user service method is fine, but check it out and let me know what you think!