Help Required (DESC)
If you look at the console on the general page, click on the description and click out to send a query to the server, and read the errors. The query requires an object, but this means that a stringified version ofthe object needs to be added into the query, but it cannot be JSON because it contains double quotes around the objects keys as well. I literally cant figure this out so if you could take a loot that would be appreciated.