UserObject Inheritance (READ DESC)
So there are three components in the profile section, the main component, sub-component, and sub-sub-component (stuff like the /information section), so I query the server for the userobject once at the main-component, and pass it to the desktop and mobile sub-components via props. Then the sub-sub-components can get it from their parent using this.$parent. This is to avoid querying the server on each component, but it might come off as hard to read or understand, so let me know if you think you have a better idea.