Sooper Spicy Commit (desc)
Bro this is gettin spicy, Now the post popup actully creates a new post in the database! It would be helpfull if you could look over my probably shitty backend code. I still have to clean a bunch of stuff up but im tired for now, but more importantly, IT WORKS AFTER HOURS OF PAIN