Working Notifications! (desc)
So the notifications work as of now, if a user likes a post, it sends a notification to the author of that post. The only issue im finding right now is that if the user unlikes and relikes the post, the notification is sent again and causes a duplicate key error. So we need to cycle through the notifcations and not allow the users to just unlike and relike constantly to keep sending notifications. Also, the frontend will need some work with longer usernames