Code refactor + work on notifications page (desc)
Im going to need your help with deciding what to do here, I've gone in between a few different ideas, and im not sure what to do. Because each message has a read variable, which effects the red number after the bell icon. So at first I was thinking we make it just read all the notifications visibile when the user goes to the notifications page, but that seemed weird, so now I tried making it so that you click on a notification to make it read. But this seems kinda overkill. Let me know what you think and how you think we should proceed from here.