Im a f*** genius
I created a general loadMoreModule that you can use with any query so I don't have to keep rewirting the logic for every scenario, its also a hell of a lot more readable than the older method. I just need to migrate all the other queries to actully use this, and well be cutting out like 500 lines of code.